requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
Django==4.0.5
django-safedelete==1.2.2
djangorestframework==3.13.1
idna==3.3
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
